Move fuzz encoding utils to own module
