+++ fuzz/fuzz_targets/entry_encoding_random.rs
@@ -45,7 +45,7 @@ impl Decodable for FakePayloadDigest {
 impl PayloadDigest for FakePayloadDigest {}
 
 use libfuzzer_sys::fuzz_target;
-use willow_data_model_fuzz::encoding_random;
+use willow_data_model_fuzz::encode::encoding_random;
 
 fuzz_target!(|data: &[u8]| {
     smol::block_on(async {
