+++ fuzz/fuzz_targets/path_encoding.rs
@@ -3,7 +3,7 @@
 use libfuzzer_sys::fuzz_target;
 use ufotofu::local_nb::consumer::TestConsumer;
 use willow_data_model::path::Path;
-use willow_data_model_fuzz::encoding_roundtrip;
+use willow_data_model_fuzz::encode::encoding_roundtrip;
 
 const MCL: usize = 111111;
 const MCC: usize = 111111;
