+++ fuzz/fuzz_targets/path_rel_path_encoding.rs
@@ -4,7 +4,7 @@ use libfuzzer_sys::fuzz_target;
 use ufotofu::local_nb::consumer::TestConsumer;
 
 use willow_data_model::path::Path;
-use willow_data_model_fuzz::relative_encoding_roundtrip;
+use willow_data_model_fuzz::encode::relative_encoding_roundtrip;
 
 const MCL: usize = 16;
 const MCC: usize = 16;
