+++ fuzz/fuzz_targets/u8be_encoding.rs
@@ -3,7 +3,7 @@
 use libfuzzer_sys::fuzz_target;
 use ufotofu::local_nb::consumer::TestConsumer;
 use willow_data_model::encoding::unsigned_int::U8BE;
-use willow_data_model_fuzz::encoding_roundtrip;
+use willow_data_model_fuzz::encode::encoding_roundtrip;
 
 fuzz_target!(|data: (u8, TestConsumer<u8, u16, ()>)| {
     let (n, mut consumer) = data;
