+++ fuzz/fuzz_targets/u8be_encoding_random.rs
@@ -2,7 +2,7 @@
 
 use libfuzzer_sys::fuzz_target;
 use willow_data_model::encoding::unsigned_int::U8BE;
-use willow_data_model_fuzz::encoding_random;
+use willow_data_model_fuzz::encode::encoding_random;
 
 fuzz_target!(|data: &[u8]| {
     smol::block_on(async {
